refactor: seperate notification strategy and manager from excavations
