+++ coral/functions/notify_excavation.py
@@ -1,9 +1,7 @@
 from arches.app.functions.base import BaseFunction
-from arches.app.models import models
-from arches.app.models.resource import Resource
 from arches_orm.adapter import admin
-import logging
-import pdb
+from coral.functions.notifications.notification_base_strategy import NotificationStrategy
+from coral.functions.notifications.notification_manager import NotificationManager
 
 # node groups
 SYSTEM_REFERENCE = '991c3c74-48b6-11ee-85af-0242ac140007'
@@ -59,80 +57,25 @@ class NotifyExcavation(BaseFunction):
         with admin():
             if context and context.get('escape_function', False):
                 return
+            
+            strategy_registry = {
+                SYSTEM_REFERENCE : SystemReferenceStrategy, 
+                REPORT : ReportStrategy, 
+                CUR_D_DECISION : CurDDecisionStrategy, 
+                CUR_E_DECISON : CurEDecisionStrategy, 
+                APPLICATION_DETAILS : ApplicationDetailsStrategy, 
+                TRANSFER_OF_LICENCE : TransferOfLicenceStrategy,
+                EXTENSION_OF_LICENCE: ExtensionOfLicenceStrategy
+            }
 
             nodegroup_id = str(tile.nodegroup_id)
 
             user = request.user
             user_resource = Person.where(user_account=user.id)[0] if user and Person.where(user_account=user.id) else None
             
-            notification_manager = NotificationManager(nodegroup_id, user_resource, tile)
+            notification_manager = NotificationManager(nodegroup_id, user_resource, tile, strategy_registry)
 
             notification_manager.notify()
-
-class NotificationStrategy:
-    def send_notification(self, user, tile):
-        raise NotImplementedError("Subclasses should implement this method")
-
-    def get_resource_details(self, tile, remove_prefix=None, remove_suffix=None):
-        resource_instance_id = str(tile.resourceinstance.resourceinstanceid)
-        name = Resource.objects.get(pk=resource_instance_id).displayname()
-        if remove_prefix:
-            name = name.removeprefix(remove_prefix).strip()
-        if remove_suffix:
-            name = name.removesuffix(remove_suffix).strip()
-        return  name, resource_instance_id
-    
-    def create_notification(self, message, name, resource_instance_id, response_slug):
-        self._delete_existing_notification(resource_instance_id)
-        notification = models.Notification(
-                    message = message,
-                    context = {
-                        "resource_instance_id": resource_instance_id,
-                        "resource_id": name,
-                        "response_slug": response_slug
-                    },
-                )
-                        
-        notification.save()
-        return notification
-    
-    def notify_groups(self, user, group_list, notification):
-        from arches_orm.models import Group, Person
-        with admin():
-            for group_id in group_list:
-                group = Group.find(group_id)
-                person_list = [Person.find(member.id) for member in group.members if isinstance(member, Person)]
-                
-                for person in person_list:
-                    if user is None or person != user:
-                        
-                        self.notify_user(person, notification)
-        
-    def notify_user(self, user, notification):
-        user = user.user_account
-
-        user_x_notification = models.UserXNotification(
-            notif = notification,
-            recipient = user
-        )
-        user_x_notification.save()
-    
-    def get_domain_value_string(self, value_id, node_id):
-        node = models.Node.objects.filter(
-            pk = node_id,   
-        ).first()
-        options = node.config.get("options")
-        value_string = next((option.get("text").get("en") for option in options if option.get("id") == value_id), None)
-        return value_string
-
-    def _delete_existing_notification(self, resource_instance_id):
-        existing_notification = models.Notification.objects.filter(
-            context__resource_instance_id=resource_instance_id
-        ).first()
-
-        if existing_notification:
-            existing_notification.delete()
-            logging.info("Deleted Existing Notification")
 
 class SystemReferenceStrategy(NotificationStrategy):
     def send_notification(self, user, tile):
@@ -273,32 +216,4 @@ class ExtensionOfLicenceStrategy(NotificationStrategy):
         notification = self.create_notification(message, name, resource_instance_id, EXCAVATION_SLUG)
                                     
         self.notify_groups(user, groups_to_notify, notification)
-
-class NotificationManager():
-    strategy_registry = {
-        SYSTEM_REFERENCE : SystemReferenceStrategy, 
-        REPORT : ReportStrategy, 
-        CUR_D_DECISION : CurDDecisionStrategy, 
-        CUR_E_DECISON : CurEDecisionStrategy, 
-        APPLICATION_DETAILS : ApplicationDetailsStrategy, 
-        TRANSFER_OF_LICENCE : TransferOfLicenceStrategy,
-        EXTENSION_OF_LICENCE: ExtensionOfLicenceStrategy
-    }
-
-    def __init__(self, node_group_id, user, tile):
-        self.user = user
-        self.node_group_id = node_group_id
-        self.tile = tile
-        self.strategy = self._select_strategy(node_group_id)
-        # self.strategy.delete_existing_notification(tile)
-
-    def _select_strategy(self, node_group_id):
-        strategy_class = self.strategy_registry.get(node_group_id)
-        if strategy_class:
-            return strategy_class()
-        else:
-            raise ValueError(f"No strategy found for node group id: {node_group_id}")
-        
-    def notify(self):
-        self.strategy.send_notification(self.user, self.tile)
